Adjusted tests to changes
